requirements.txt
click==8.1.3
colorama==0.4.6
Faker==16.4.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2